rework components that came from partials
