+++ ui/lib/replication/addon/components/replication-actions-single.js
@@ -1,0 +1,14 @@
+import Component from '@ember/component';
+
+export default Component.extend({
+  onSubmit() {},
+  replicationMode: null,
+  replicationDisplayMode: null,
+  model: null,
+
+  actions: {
+    onSubmit() {
+      return this.onSubmit(...arguments);
+    },
+  },
+});
