+++ ui/lib/replication/addon/controllers/replication-mode.js
@@ -5,4 +5,12 @@ import Controller from '@ember/controller';
 export default Controller.extend({
   rm: service('replication-mode'),
   replicationMode: alias('rm.mode'),
+  actions: {
+    onEnable(mode) {
+      return this.transitionToRoute('mode', mode);
+    },
+    onDisable() {
+      return this.transitionToRoute('index');
+    },
+  },
 });
