+++ ui/lib/replication/addon/templates/components/replication-actions.hbs
@@ -1,5 +1,17 @@
-{{message-error errors=errors}}
 {{#if loading}}
+  <LayoutLoading />
 {{else}}
-  {{partial (concat 'partials/replication/' selectedAction)}}
+  <MessageError @errors={{errors}} />
+  {{#each (replication-action-for-mode replicationMode model.replicationAttrs.modeForUrl) as |replicationAction index|}}
+    {{#if (get model (concat 'can' (camelize replicationAction)))}}
+      <div class="box is-fullwidth is-marginless {{if (gt index 0) 'is-bottomless' 'is-shadowless'}}">
+        {{component (concat 'replication-action-' replicationAction)
+          onSubmit=(action "onSubmit")
+          replicationMode=replicationMode
+          model=model
+          replicationDisplayMode=replicationDisplayMode
+        }}
+      </div>
+    {{/if}}
+  {{/each}}
 {{/if}}
