+++ ui/lib/replication/addon/templates/index.hbs
@@ -13,5 +13,7 @@
   <ReplicationSummary
     @cluster={{model}}
     @showModeSummary={{true}}
+    @onEnable={{action "onEnable"}}
+    @onDisable={{action "onDisable"}}
   />
 {{/if}}
